requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
asgiref==3.4.1
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
base58==2.1.0
bleach==4.0.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
cycler==0.10.0
dataclasses==0.8
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.6
django-picklefield==3.0.1
django-utils-six==2.0
django-widget-tweaks==1.4.8
dominate==2.6.0
EasyProcess==0.3
entrypoint2==0.2.4
entrypoints==0.3
filelock==3.0.12
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
gitdb==4.0.7
GitPython==3.1.18
idna==3.2
importlib-metadata==4.6.4
importlib-resources==5.2.2
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==7.0.1
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mss==6.1.0
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.19.5
packaging==21.0
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
pipenv==2021.5.29
platformdirs==2.2.0
prettytable==2.1.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
pyarrow==5.0.0
pycparser==2.20
pydeck==0.6.2
pygi==1.0.3
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
pyscreenshot==1.0
python-dateutil==2.8.2
python-decouple==3.4
python-Levenshtein==0.12.2
pytz==2021.1
pywin32==301
pywinpty==1.1.3
pyzmq==22.2.1
requests==2.26.0
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
smmap==4.0.0
sqlparse==0.4.1
streamlit==0.87.0
terminado==0.11.1
testpath==0.5.0
threadpoolctl==2.2.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==4.3.3
typing-extensions==3.10.0.0
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.6
validators==0.18.2
virtualenv==20.7.2
virtualenv-clone==0.5.6
visitor==0.1.3
watchdog==2.1.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
zipp==3.5.0
